requirements.txt
setuptools
Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
supabase==2.3.0
bcrypt==4.1.2
razorpay==1.4.1
gunicorn==21.2.0